add: add ra-boilerplate

+++ clean.sh
@@ -1,9 +1,15 @@
 #!/bin/sh
 
-echo "START | Clean up SEMS_FE project..."
+echo "START | Clean up RA-Boilerplate project..."
 
-rm -rf build/*
+folder_path="./build"
+
+if [ -d "$folder_path" ]; then
+    echo "Folder existed!!"
+else
+  rm -rf build/*
+fi
+
 echo "Removed build"
 
-echo "DONE | Clean up SEMS_FE project..."
-
+echo "DONE | Clean up RA-Boilerplate project..."
